handlers: deal with redirected WDT reply correctly
In WDT reply handler previous context data are referenced
in case task is redirected. This is not a correct behavior,
previous context must not be referenced as there are no
guarantees data are still there.

This commit saves forwarder address into task struture
and uses it to match forwarder in WDT reply handler in case
task is redirected.

Signed-off-by: Oleksandr Natalenko <oleksandr@redhat.com>